carrier_verification/requirements.txt
fastapi==0.99.0
mangum==0.19.0
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
exceptiongroup==1.1.1
typing-extensions==4.6.0
SQLAlchemy==2.0.23
pydantic>=1.7.4,<2.0.0
psycopg2-binary==2.9.9  # Changed from psycopg2 to psycopg2-binary